configs-service/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
pydantic
